remove useless casts

+++ ww/base/wlog.c
@@ -598,14 +598,14 @@ void loggerDestroyDefaultLogger(void)
 void stdoutLogger(int loglevel, const char *buf, int len)
 {
     (void) loglevel;
-    ssize_t count = write(fileno(stdout), buf, (size_t) (len));
+    ssize_t count = write(fileno(stdout), buf,  len);
     (void) count;
 }
 
 void stderrLogger(int loglevel, const char *buf, int len)
 {
     (void) loglevel;
-    ssize_t count = write(fileno(stderr), buf, (size_t) (len));
+    ssize_t count = write(fileno(stderr), buf, len);
     (void) count;
 }
